fix: dialog folder architecture
