--- packages/app/src/components/index.ts
@@ -1,1 +1,0 @@
-export { default as Portal } from './Portal'
