--- packages/app/src/providers/index.ts
@@ -1,1 +1,0 @@
-export { default as DialogProvider } from './DialogProvider'
